config/requirements.txt
# =============================================================================
# Bubble - Blockchain Analytics Platform
# Copyright (c) 2025-2026 All Rights Reserved.
# =============================================================================
#
# Python 3.12 Dependencies
# =============================================================================

# Web Framework
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0

# GraphQL
graphene>=3.0.0
graphene-sqlalchemy>=3.0.0rc2
graphql-relay
graphql-server-core

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Web3 & Blockchain
web3>=6.0.0

# HTTP Requests
requests>=2.31.0

# TigerGraph
pyTigerGraph>=1.0.0

# Environment Variables
python-dotenv>=1.0.0

# Date & Time
python-dateutil>=2.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Testing (like mission7)
pytest>=8.0.0
pytest-cov>=4.0.0

